Move HTTP integration tests into a separate file
